Updated theory documentaition
