+++ konstraints-core/src/main/kotlin/tools/aqua/konstraints/smt/ArraysEx.kt
@@ -20,10 +20,15 @@ package tools.aqua.konstraints.smt
 
 import tools.aqua.konstraints.parser.*
 
+/*
+ * This file implements the SMT ArraysEx theory
+ * https://smt-lib.org/theories-ArraysEx.shtml
+ */
+
 /**
  * Array selection operation.
- * - (par (X Y) (select (Array X Y) X Y)
- * - (select [array] [index])
+ * - `(par (X Y) (select (Array X Y) X Y)`
+ * - `(select [array] [index])`
  */
 class ArraySelect<X : Sort, Y : Sort>(
     val array: Expression<ArraySort<X, Y>>,
@@ -46,8 +51,8 @@ class ArraySelect<X : Sort, Y : Sort>(
 
 /**
  * Array store operation.
- * - (par (X Y) (store (Array X Y) X Y (Array X Y)))
- * - (store [array] [index] [value]).
+ * - `(par (X Y) (store (Array X Y) X Y (Array X Y)))`
+ * - `(store [array] [index] [value])`
  */
 class ArrayStore<X : Sort, Y : Sort>(
     val array: Expression<ArraySort<X, Y>>,
