+++ konstraints-core/src/main/kotlin/tools/aqua/konstraints/smt/Reals_Ints.kt
@@ -20,10 +20,15 @@ package tools.aqua.konstraints.smt
 
 import tools.aqua.konstraints.parser.*
 
+/*
+ * This file implements the SMT Reals_Ints theory
+ * https://smt-lib.org/theories-Reals_Ints.shtml
+ */
+
 /**
  * Convert an integer [inner] to real.
- * - (to_real Int Real)
- * - (to_real [inner])
+ * - `(to_real Int Real)`
+ * - `(to_real [inner])`
  */
 class ToReal(override val inner: Expression<IntSort>) :
     UnaryExpression<RealSort, IntSort>("to_real".toSymbolWithQuotes(), Real) {
@@ -35,8 +40,8 @@ class ToReal(override val inner: Expression<IntSort>) :
 
 /**
  * Convert a real [inner] to int.
- * - (to_int Real Int)
- * - (to_int [inner])
+ * - `(to_int Real Int)`
+ * - `(to_int [inner])`
  */
 class ToInt(override val inner: Expression<RealSort>) :
     UnaryExpression<IntSort, RealSort>("to_int".toSymbolWithQuotes(), SMTInt) {
@@ -46,7 +51,7 @@ class ToInt(override val inner: Expression<RealSort>) :
       ToIntDecl.constructDynamic(children, emptyList())
 }
 
-/** (is_int Real Bool). */
+/** `(is_int Real Bool)`. */
 class IsInt(override val inner: Expression<RealSort>) :
     UnaryExpression<BoolSort, RealSort>("is_int".toSymbolWithQuotes(), Bool) {
   override val theories = REALS_INTS_MARKER_SET
